+++ package.json
@@ -16,6 +16,8 @@
         "request": "^2.88.2"
     },
     "devDependencies": {
-        "electron": "^32.1.2"
+        "@types/node": "^22.7.4",
+        "electron": "^32.1.2",
+        "typescript": "^5.6.2"
     }
 }
