--- src/components/js/progress-bar.js
@@ -1,6 +1,0 @@
-const progressBar = document.getElementById("bar");
-
-// Just a function for the bar to be set to the passed width. ¯\_(ツ)_/¯
-function setBar(percentage) {
-    progressBar.style.width = percentage + "%";
-}
