+++ src/css/license.css
