+++ src/flashing.html
@@ -3,8 +3,8 @@
     <head>
         <meta charset="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="stylesheet" href="./components/css/main.css" />
-        <link rel="stylesheet" href="./components/css/flashing.css" />
+        <link rel="stylesheet" href="./css/main.css" />
+        <link rel="stylesheet" href="./css/flashing.css" />
         <!-- FONT -->
         <link rel="preconnect" href="https://fonts.googleapis.com" />
         <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
@@ -23,20 +23,16 @@
             <div class="buttons" id="buttons">
                 <!-- Minimise button -->
                 <button id="minBtn" href="" class="btn">
-                    <img src="./components/images/minimize.svg" height="15px" />
+                    <img src="./images/minimize.svg" height="15px" />
                 </button>
                 <!-- Close button -->
                 <button id="closeBtn" href="" class="btn">
-                    <img src="./components/images/close.svg" height="16px" />
+                    <img src="./images/close.svg" height="16px" />
                 </button>
             </div>
             <!-- Window Title -->
             <div class="title" id="title">
-                <img
-                    src="./components/images/icon.png"
-                    height="20px"
-                    class="titlelogo"
-                />
+                <img src="./images/icon.png" height="20px" class="titlelogo" />
                 &nbsp; &nbsp;
                 <p>AvdanOS Imager</p>
             </div>
@@ -57,8 +53,8 @@
             </div>
         </main>
         <footer></footer>
-        <script src="./components/js/renderer.js"></script>
+        <script src="./js/renderer.js"></script>
         <!-- Discord RPC Region, please put all scripts above -->
-        <script src="./components/js/RPC/flashing.js"></script>
+        <script src="./js/RPC/flashing.js"></script>
     </body>
 </html>
