+++ src/license.html
@@ -11,11 +11,7 @@
             type="application/rdf+xml"
             href="http://www.gnu.org/licenses/gpl-3.0.rdf"
         />
-        <link
-            rel="stylesheet"
-            href="./components/css/license.css"
-            type="text/css"
-        />
+        <link rel="stylesheet" href="./css/license.css" type="text/css" />
     </head>
     <body>
         <h3 style="text-align: center">GNU GENERAL PUBLIC LICENSE</h3>
